# Another tiny speedup to Day 11