Before:

```
|    Method | IsPartOne | Day |         Mean |      Error |     StdDev |
|---------- |---------- |---- |-------------:|-----------:|-----------:|
| Benchmark |     False |  11 |     7.353 us |  0.0549 us |  0.0514 us |
| Benchmark |      True |  11 | 2,868.372 us | 33.0022 us | 29.2556 us |
```

After:

```
|    Method | IsPartOne | Day |         Mean |     Error |    StdDev |
|---------- |---------- |---- |-------------:|----------:|----------:|
| Benchmark |     False |  11 |     7.244 us | 0.0219 us | 0.0183 us |
| Benchmark |      True |  11 | 2,839.408 us | 9.6564 us | 8.5601 us |
```